Fix enabling multiple languages if the `APP_NAME` in the db was "Arches" #11660 (#11687)
* Use same guard for "last" project as project-level urls.py #11660

This check sniffs for whether this urls.py is the one being run as a project
(not included by any further apps). For core arches, this would only be true
if running core arches without a project. This was failing on a Django exception
inside i18n_patterns().

This check is the same check that we're using in project-level settings.py

* Also avoid checking APP_NAME in generate_frontend_configuration()